% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter1-foundations.R
\name{acg_loglik}
\alias{acg_loglik}
\title{Angular central Gaussian log-likelihood}
\usage{
acg_loglik(x, shape, include_constant = FALSE)
}
\arguments{
\item{x}{Nonzero directions in rows.}

\item{shape}{A positive definite shape matrix.}

\item{include_constant}{Include the density normalizing constant.}
}
\value{
A scalar log-likelihood.
}
\description{
Evaluates the angular central Gaussian log-likelihood for nonzero row
vectors. Row lengths need not equal one because they cancel from the
quadratic contribution.
}
\examples{
directions <- rbind(c(1, 0), c(0, 1), c(1, 1))
acg_loglik(directions, diag(2))
}
\references{
Tyler, D. E. (1987). A distribution-free M-estimator of multivariate
scatter. \emph{Annals of Statistics}, 15, 234--251.
\doi{10.1214/aos/1176350263}.
}
