% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ch4_cp_ordered_pair_square_sum_cpp}
\alias{ch4_cp_ordered_pair_square_sum_cpp}
\title{Ordered-pair squared inner-product sum}
\usage{
ch4_cp_ordered_pair_square_sum_cpp(signs)
}
\arguments{
\item{signs}{Matrix with observations in rows.}
}
\value{
\verb{sum_\{i != j\} (u_i' u_j)^2}.
}
\description{
Ordered-pair squared inner-product sum
}
\keyword{internal}
