% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_bai_saranadasa_two_sample}
\alias{cpp_bai_saranadasa_two_sample}
\title{Original Bai--Saranadasa two-sample statistic kernel}
\usage{
cpp_bai_saranadasa_two_sample(x, y)
}
\arguments{
\item{x, y}{Numeric observation-by-variable matrices.}
}
\value{
Internal list of statistic components.
}
\description{
Original Bai--Saranadasa two-sample statistic kernel
}
\keyword{internal}
