% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_ch4_afc_spatial_kendall}
\alias{cpp_ch4_afc_spatial_kendall}
\title{Spatial Kendall matrix for robust latent-factor extraction}
\usage{
cpp_ch4_afc_spatial_kendall(x)
}
\arguments{
\item{x}{Finite observation-by-variable matrix.}
}
\value{
Spatial Kendall matrix and the number of zero pair differences.
}
\description{
Spatial Kendall matrix for robust latent-factor extraction
}
\keyword{internal}
