% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_ch4_completion_weighted_alpha_q}
\alias{cpp_ch4_completion_weighted_alpha_q}
\title{Weighted spatial-sign alpha quadratic form}
\usage{
cpp_ch4_completion_weighted_alpha_q(directions, h, weights)
}
\arguments{
\item{directions}{Observation-by-asset spatial-sign matrix.}

\item{h}{Residualized-intercept vector.}

\item{weights}{Evaluated radial weights.}
}
\value{
The quadratic form and empirical second weight moment.
}
\description{
Weighted spatial-sign alpha quadratic form
}
\keyword{internal}
