% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_ch7sc_feature_scores}
\alias{cpp_ch7sc_feature_scores}
\title{Across-center feature scores for Sparse--SM}
\usage{
cpp_ch7sc_feature_scores(centers)
}
\arguments{
\item{centers}{Finite cluster-by-variable center matrix.}
}
\value{
Internal center average and hard-screening scores.
}
\description{
Across-center feature scores for Sparse--SM
}
\keyword{internal}
