% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_chen_qin_two_sample}
\alias{cpp_chen_qin_two_sample}
\title{Chen--Qin two-sample statistic kernel}
\usage{
cpp_chen_qin_two_sample(x, y)
}
\arguments{
\item{x, y}{Numeric observation-by-variable matrices.}
}
\value{
Internal list of statistic components.
}
\description{
Chen--Qin two-sample statistic kernel
}
\keyword{internal}
