% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_clx_two_sample}
\alias{cpp_clx_two_sample}
\title{Oracle or feasible Cai--Liu--Xia statistic kernel}
\usage{
cpp_clx_two_sample(x, y, precision, oracle)
}
\arguments{
\item{x, y}{Numeric observation-by-variable matrices.}

\item{precision}{Symmetric precision or estimated-precision matrix.}

\item{oracle}{If true, use diag(precision); otherwise use equation (7).}
}
\value{
Internal list of statistic components.
}
\description{
Oracle or feasible Cai--Liu--Xia statistic kernel
}
\keyword{internal}
