% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_fzwz_bf_two_sample}
\alias{cpp_fzwz_bf_two_sample}
\title{Feng--Zou--Wang--Zhu Behrens--Fisher statistic kernel}
\usage{
cpp_fzwz_bf_two_sample(x, y)
}
\arguments{
\item{x, y}{Numeric observation-by-variable matrices.}
}
\value{
Internal list of statistic components.
}
\description{
Feng--Zou--Wang--Zhu Behrens--Fisher statistic kernel
}
\keyword{internal}
