% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_skk_two_sample}
\alias{cpp_skk_two_sample}
\title{Srivastava--Katayama--Kano statistic kernel}
\usage{
cpp_skk_two_sample(x, y)
}
\arguments{
\item{x, y}{Numeric observation-by-variable matrices.}
}
\value{
Internal list of statistic components.
}
\description{
Srivastava--Katayama--Kano statistic kernel
}
\keyword{internal}
