% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_srivastava_du_one_sample}
\alias{cpp_srivastava_du_one_sample}
\title{Srivastava--Du one-sample statistic kernel}
\usage{
cpp_srivastava_du_one_sample(x, mu)
}
\arguments{
\item{x}{Numeric observation-by-variable matrix.}

\item{mu}{Numeric null-mean vector.}
}
\value{
Internal list of statistic components.
}
\description{
Srivastava--Du one-sample statistic kernel
}
\keyword{internal}
