% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_wang_peng_li_one_sample}
\alias{cpp_wang_peng_li_one_sample}
\title{Wang--Peng--Li one-sample spatial-sign statistic kernel}
\usage{
cpp_wang_peng_li_one_sample(x, mu)
}
\arguments{
\item{x}{Numeric observation-by-variable matrix.}

\item{mu}{Numeric null-location vector.}
}
\value{
Internal list of statistic and cross-validation components.
}
\description{
Wang--Peng--Li one-sample spatial-sign statistic kernel
}
\keyword{internal}
