% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_zhang_feng_two_sample_scores}
\alias{cpp_zhang_feng_two_sample_scores}
\title{Two-sample Zhang--Feng marginal WMW scores}
\usage{
cpp_zhang_feng_two_sample_scores(x, y)
}
\arguments{
\item{x}{First numeric observation-by-variable matrix.}

\item{y}{Second numeric observation-by-variable matrix.}
}
\value{
Internal list of marginal scores and exact untied null moments.
}
\description{
Two-sample Zhang--Feng marginal WMW scores
}
\keyword{internal}
