% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDElliptical-package.R
\docType{package}
\name{HDElliptical-package}
\alias{HDElliptical}
\alias{HDElliptical-package}
\title{HDElliptical: high-dimensional elliptical methods}
\description{
\code{HDElliptical} provides R and Rcpp implementations accompanying
\emph{High-Dimensional Data Analysis for Elliptically Symmetric Distributions}.
Functions use observations in rows and variables in columns throughout.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/flnankai/HDElliptical}
  \item Report bugs at \url{https://github.com/flnankai/HDElliptical/issues}
}

}
\author{
\strong{Maintainer}: Long Feng \email{flnankai@nankai.edu.cn} (Copyright holder for the HDElliptical implementation) [copyright holder]

Other contributors:
\itemize{
  \item Dan Zhuang (SEMC methodology and the external GEMcluster reference implementation used for validation; no GEMcluster source code is included) [contributor]
}

}
\keyword{internal}
